requirements.txt
absl-py==2.1.0
alembic==1.13.2
asgiref==3.8.1
astunparse==1.6.3
bcrypt==4.1.3
blinker==1.8.2
Brotli==1.1.0
certifi==2021.10.8
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
Django==5.1.5
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.1
Flask==3.0.3
Flask-Alchemist==1.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-LoginManager==1.1.6
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.6.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.4.0
protobuf==4.25.3
psutil==6.0.0
pycparser==2.22
pydyf==0.11.0
pygame-ce==2.5.1
Pygments==2.18.0
pyparsing==3.1.2
pyphen==0.15.0
python-dateutil==2.9.0.post0
requests==2.31.0
rich==13.7.1
scapy==2.5.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.3
sympy==1.12.1
talisman==0.1.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==2.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.11.0
urllib3==2.2.1
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
zopfli==0.2.3.post1
